Log the hell out of Delta creation and application

+++ state/delta.go
@@ -41,10 +41,14 @@ func (session *Session) Between(desired *DesiredState, actual *ActualState) Delt
 	)
 
 	for filePath, desiredContent := range desired.Files {
+		log.WithField("filePath", filePath).Debug("Verifying expected file.")
 		actualContent, ok := actual.Files[filePath]
 		if !ok || !bytes.Equal(desiredContent, actualContent) {
 			filesToWrite = append(filesToWrite, filePath)
 			fileContentByPath[filePath] = desiredContent
+			log.WithField("filePath", filePath).Debug("File was absent or different.")
+		} else {
+			log.WithField("filePath", filePath).Debug("Nothing to do.")
 		}
 	}
 
@@ -55,6 +59,7 @@ func (session *Session) Between(desired *DesiredState, actual *ActualState) Delt
 
 	for _, actual := range actual.Units {
 		if desired, ok := desiredByName[actual.UnitName()]; ok {
+			log.WithField("unitName", actual.UnitName()).Debug("Verifying systemd unit.")
 			desiredRemaining[desired.UnitName()] = false
 
 			// Determine if the actual unit needs to be reloaded to match the desired one.
@@ -68,6 +73,7 @@ func (session *Session) Between(desired *DesiredState, actual *ActualState) Delt
 				continue
 			}
 			if !bytes.Equal(expected.Bytes(), actual.Content) {
+				log.WithField("unitName", actual.UnitName()).Debug("Unit content differs.")
 				unitsToChange = append(unitsToChange, desired)
 				continue
 			}
@@ -83,6 +89,10 @@ func (session *Session) Between(desired *DesiredState, actual *ActualState) Delt
 			if err != nil {
 				if client.IsErrNotFound(err) {
 					// It's not running. Definitely restart the thing.
+					log.WithFields(log.Fields{
+						"unitName":      actual.UnitName(),
+						"containerName": desired.Container.Name,
+					}).Debug("Container is not running.")
 					unitsToRestart = append(unitsToRestart, desired)
 					continue
 				}
@@ -92,6 +102,12 @@ func (session *Session) Between(desired *DesiredState, actual *ActualState) Delt
 
 			if container.Image != desired.Container.ImageID {
 				// A newer image has been pulled. Restart the unit to pick it up.
+				log.WithFields(log.Fields{
+					"unitName":       actual.UnitName(),
+					"containerName":  desired.Container.Name,
+					"desiredImageID": desired.Container.ImageID,
+					"actualImageID":  container.Image,
+				}).Debug("Container image is out of date.")
 				unitsToRestart = append(unitsToRestart, desired)
 				continue
 			}
@@ -100,14 +116,20 @@ func (session *Session) Between(desired *DesiredState, actual *ActualState) Delt
 			for hostPath := range desired.Volumes {
 				if _, ok := fileContentByPath[hostPath]; ok {
 					// A mounted file has been written. Restart the unit to pick it up.
+					log.WithFields(log.Fields{
+						"unitName":        actual.UnitName(),
+						"mountedFilePath": hostPath,
+					}).Debug("Mounted volume file has been changed.")
 					unitsToRestart = append(unitsToRestart, desired)
 					break
 				}
 			}
 
 			// Otherwise: everything is fine, nothing to do.
+			log.WithField("unitName", actual.UnitName()).Debug("Nothing to do.")
 		} else {
 			// Unit is no longer desired.
+			log.WithField("unitName", actual.UnitName()).Debug("Unit is no longer desired.")
 			unitsToRemove = append(unitsToRemove, actual)
 		}
 	}
@@ -116,6 +138,7 @@ func (session *Session) Between(desired *DesiredState, actual *ActualState) Delt
 	for desiredName, remaining := range desiredRemaining {
 		if remaining {
 			if desired, ok := desiredByName[desiredName]; ok {
+				log.WithField("unitName", desired.UnitName()).Debug("Unit is not yet present.")
 				unitsToAdd = append(unitsToAdd, desired)
 			}
 		}
@@ -154,6 +177,8 @@ func (d Delta) Apply() []error {
 			errs = append(errs, err)
 			continue
 		}
+
+		log.WithField("filePath", filePath).Info("File content written.")
 	}
 
 	for _, unit := range d.UnitsToAdd {
@@ -165,6 +190,11 @@ func (d Delta) Apply() []error {
 
 		errs = append(errs, session.WriteUnit(unit, f)...)
 		f.Close()
+
+		log.WithFields(log.Fields{
+			"unitName":     unit.UnitName(),
+			"unitFilePath": unit.Path,
+		}).Info("Unit file created.")
 	}
 
 	for _, unit := range d.UnitsToChange {
@@ -179,6 +209,11 @@ func (d Delta) Apply() []error {
 
 		errs = append(errs, d.session.WriteUnit(unit, f)...)
 		f.Close()
+
+		log.WithFields(log.Fields{
+			"unitName":     unit.UnitName(),
+			"unitFilePath": unit.Path,
+		}).Info("Unit file modified.")
 	}
 
 	for _, unit := range d.UnitsToRestart {
@@ -192,53 +227,100 @@ func (d Delta) Apply() []error {
 		for _, unit := range d.UnitsToRemove {
 			disablePaths = append(disablePaths, unit.Path)
 
+			log.WithField("unitName", unit.UnitName()).Debug("Stopping unit.")
 			if _, err := session.conn.StopUnit(unit.UnitName(), "replace", stops); err != nil {
 				errs = append(errs, fmt.Errorf("Unable to stop unit %s (%v)", unit.UnitName(), err))
+				stops <- ""
+
+				log.WithField("unitName", unit.UnitName()).Info("Killing unit.")
 				session.conn.KillUnit(unit.Path, 9)
+				log.WithField("unitName", unit.UnitName()).Info("Unit killed.")
 			}
 		}
 		for i := 0; i < len(d.UnitsToRemove); i++ {
 			<-stops
 		}
+		log.WithField("count", len(d.UnitsToRemove)).Debug("Units stopped or killed.")
+
+		log.WithField("unitPaths", disablePaths).Debug("Disabling units.")
 		if _, err := session.conn.DisableUnitFiles(disablePaths, false); err != nil {
 			errs = append(errs, fmt.Errorf("Unable to disable units %v (%v)", disablePaths, err))
 		}
+		log.WithField("count", len(disablePaths)).Debug("Units disabled.")
+	} else {
+		log.Debug("No units to remove.")
 	}
 
 	// Reload to pick up any rewritten unit files.
 	if needsReload {
+		log.Debug("Reloading systemd unit files.")
 		if err := session.conn.Reload(); err != nil {
 			errs = append(errs, err)
 			return errs
 		}
+		log.Debug("Reloaded successfully.")
 	}
 
 	// Start and enable newly created units.
-	starts := make(chan string, len(d.UnitsToAdd))
-	enablePaths := make([]string, 0, len(d.UnitsToAdd))
-	for _, unit := range d.UnitsToAdd {
-		enablePaths = append(enablePaths, unit.Path)
-		if _, err := session.conn.StartUnit(unit.UnitName(), "replace", starts); err != nil {
-			errs = append(errs, fmt.Errorf("Unable to start unit %s (%v)", unit.UnitName(), err))
+	if len(d.UnitsToAdd) > 0 {
+		log.WithField("count", len(d.UnitsToAdd)).Debug("Starting and enabling units.")
+
+		starts := make(chan string, len(d.UnitsToAdd))
+		enablePaths := make([]string, 0, len(d.UnitsToAdd))
+		for _, unit := range d.UnitsToAdd {
+			enablePaths = append(enablePaths, unit.Path)
+			log.WithField("unitName", unit.UnitName()).Debug("Starting unit.")
+			if _, err := session.conn.StartUnit(unit.UnitName(), "replace", starts); err != nil {
+				errs = append(errs, fmt.Errorf("Unable to start unit %s (%v)", unit.UnitName(), err))
+				starts <- ""
+			}
 		}
+		for i := 0; i < len(d.UnitsToAdd); i++ {
+			<-starts
+		}
+		log.WithField("count", len(d.UnitsToAdd)).Info("Units started.")
+
+		log.WithField("count", len(enablePaths)).Info("Enabling units.")
+		if _, _, err := session.conn.EnableUnitFiles(enablePaths, false, true); err != nil {
+			errs = append(errs, fmt.Errorf("Unable to enable units %v (%v)", enablePaths, err))
+		}
+		log.WithField("count", len(enablePaths)).Debug("Units enabled.")
+	} else {
+		log.Debug("No units to start and enable.")
 	}
-	for i := 0; i < len(d.UnitsToAdd); i++ {
-		<-starts
-	}
-	session.conn.EnableUnitFiles(enablePaths, false, true)
 
 	// Restart changed units and units whose containers have been updated.
-	restarts := make(chan string, len(restartUnits))
-	for _, unitName := range restartUnits {
-		if _, err := session.conn.RestartUnit(unitName, "replace", restarts); err != nil {
-			errs = append(errs, fmt.Errorf("Unable to restart unit %s (%v)", unitName, err))
+	if len(restartUnits) > 0 {
+		log.WithField("count", len(restartUnits)).Debug("Restarting units.")
+
+		restarts := make(chan string, len(restartUnits))
+		for _, unitName := range restartUnits {
+			log.WithField("unitName", unitName).Debug("Restarting unit.")
+			if _, err := session.conn.RestartUnit(unitName, "replace", restarts); err != nil {
+				errs = append(errs, fmt.Errorf("Unable to restart unit %s (%v)", unitName, err))
+				restarts <- ""
+			}
 		}
+
+		for i := 0; i < len(restartUnits); i++ {
+			<-restarts
+		}
+		log.WithField("count", len(restartUnits)).Info("Units restarted.")
+	} else {
+		log.Debug("No units to restart.")
 	}
 
-	for _, unit := range d.UnitsToRemove {
-		if err := os.Remove(unit.Path); err != nil {
-			errs = append(errs, fmt.Errorf("Unable to remove unit source for %s (%v)", unit.Path, err))
+	if len(d.UnitsToRemove) > 0 {
+		log.WithField("count", len(d.UnitsToRemove)).Debug("Removing unit files.")
+		for _, unit := range d.UnitsToRemove {
+			log.WithField("unitFilePath", unit.Path).Debug("Removing unit file.")
+			if err := os.Remove(unit.Path); err != nil {
+				errs = append(errs, fmt.Errorf("Unable to remove unit source for %s (%v)", unit.Path, err))
+			}
+			log.WithField("unitFilePath", unit.Path).Info("Removed unit file.")
 		}
+	} else {
+		log.Debug("No unit files to remove.")
 	}
 
 	return errs
